controller/pull: hotfix for PR-up-to-date check
We have always assumed that the commit lists returned by the GitHub API
come from git directly (e.g. with git-log). As such, we further assumed
that the ordering of commits returned by the GitHub API are such that
the parent commits are always returned first. However, this turns out
not to be the place, leading to our checkForBranchOutOfDate() to return
false negatives, resulting in users unable to submit their PR.

For now, let's put a band aid by checking every single commit. This
should work in this case since we already verify that there are no merge
commits.

However, a more long-term solution would be to topologically sort the
commit list returned by GitHub.