prstore#forceCherryPick(): do not punt on non-conflict status entries
Sometimes, even if there is a conflict in a cherry-pick, not all entries
in the index may be conflicting (the merge proceeded successfully for
those files). However, we did not expect `git-status` to return
non-conflict entries, and wrongly error-ed out on them.

Fix this by handling non-conflict entries as well.